Refactor native prop resolving logic
Consolidate logic for resolving global props into a new hook -
useNativeProps.

Simply replicating the API patterns from web is not
ideal for native, where we need to polyfill more complex behaviors like
style inheritance. This diff starts the process of customizing the prop
transforms to enable more optimizations.
