+++ packages/react-strict-dom/src/native/modules/TextString.js
@@ -9,9 +9,7 @@
 
 import * as React from 'react';
 import { Text } from 'react-native';
-import { resolveUnitlessLineHeight } from './resolveUnitlessLineHeight';
 import { useCustomProperties } from './ContextCustomProperties';
-import { useInheritedStyles } from './ContextInheritedStyles';
 import { useStyleProps } from './useStyleProps';
 
 type Props = $ReadOnly<{|
@@ -22,30 +20,16 @@ export function TextString(props: Props): React$MixedElement {
   const { children } = props;
 
   const customProperties = useCustomProperties();
-  const inheritedStyles: $FlowFixMe = useInheritedStyles();
-  const inheritedFontSize =
-    typeof inheritedStyles?.fontSize === 'number'
-      ? inheritedStyles?.fontSize
-      : undefined;
 
-  const styleProps = useStyleProps(inheritedStyles, {
+  const { nativeProps } = useStyleProps(null, {
     customProperties,
-    inheritedFontSize: inheritedFontSize
+    provideInheritableStyle: false,
+    withInheritedStyle: true,
+    withTextStyle: true
   });
-
-  if (
-    styleProps.style != null &&
-    typeof styleProps.style === 'object' &&
-    Object.keys(styleProps.style).length === 0
-  ) {
-    // $FlowFixMe (safe to remove style at this point)
-    delete styleProps.style;
-  }
-
-  resolveUnitlessLineHeight(styleProps.style);
 
   return (
     // $FlowFixMe
-    <Text {...styleProps} children={children} />
+    <Text {...nativeProps} children={children} />
   );
 }
