+++ packages/react-strict-dom/src/native/stylex/__tests__/parseTransform-test.js
@@ -8,72 +8,72 @@
 import { parseTransform } from '../parseTransform';
 
 describe('parseTransform', () => {
-  it('perspective', () => {
+  test('perspective', () => {
     expect(parseTransform('perspective(10px)')).toEqual([{ perspective: 10 }]);
   });
 
-  it('rotate', () => {
+  test('rotate', () => {
     expect(parseTransform('rotate(10deg)')).toEqual([{ rotate: '10deg' }]);
   });
 
-  it('rotateX', () => {
+  test('rotateX', () => {
     expect(parseTransform('rotateX(10deg)')).toEqual([{ rotateX: '10deg' }]);
   });
 
-  it('rotateY', () => {
+  test('rotateY', () => {
     expect(parseTransform('rotateY(10deg)')).toEqual([{ rotateY: '10deg' }]);
   });
 
-  it('rotateZ', () => {
+  test('rotateZ', () => {
     expect(parseTransform('rotateZ(10deg)')).toEqual([{ rotateZ: '10deg' }]);
   });
 
-  it('scale', () => {
+  test('scale', () => {
     expect(parseTransform('scale(3.14)')).toEqual([{ scale: 3.14 }]);
   });
 
-  it('scaleX', () => {
+  test('scaleX', () => {
     expect(parseTransform('scaleX(3.14)')).toEqual([{ scaleX: 3.14 }]);
   });
 
-  it('scaleY', () => {
+  test('scaleY', () => {
     expect(parseTransform('scaleY(3.14)')).toEqual([{ scaleY: 3.14 }]);
   });
 
-  it('scaleZ', () => {
+  test('scaleZ', () => {
     expect(parseTransform('scaleZ(3.14)')).toEqual([{ scaleZ: 3.14 }]);
   });
 
-  it('translateX', () => {
+  test('translateX', () => {
     expect(parseTransform('translateX(10px)')).toEqual([{ translateX: 10 }]);
   });
 
-  it('translateY', () => {
+  test('translateY', () => {
     expect(parseTransform('translateY(10px)')).toEqual([{ translateY: 10 }]);
   });
 
-  it('skewX', () => {
+  test('skewX', () => {
     expect(parseTransform('skewX(10deg)')).toEqual([{ skewX: '10deg' }]);
   });
 
-  it('skewY', () => {
+  test('skewY', () => {
     expect(parseTransform('skewY(10deg)')).toEqual([{ skewY: '10deg' }]);
   });
 
-  it('matrix', () => {
+  test('matrix', () => {
     expect(parseTransform('matrix(1, 2, 3, 4, 5, 6)')).toEqual([
       { matrix: [1, 2, 0, 0, 3, 4, 0, 0, 5, 6, 1, 0, 0, 0, 0, 1] }
     ]);
   });
 
-  it('multiple transforms', () => {
+  test('multiple transforms', () => {
     expect(parseTransform('translateX(10px) scaleX(1.3)')).toEqual([
       { translateX: 10 },
       { scaleX: 1.3 }
     ]);
   });
 
-  it('scientific notation', () => {
+  test('scientific notation', () => {
     expect(parseTransform('translateX(-1e3px)')).toEqual([
       { translateX: -1000 }
     ]);
@@ -82,11 +82,12 @@ describe('parseTransform', () => {
     ]);
   });
 
-  it('invalid transform', () => {
+  test('invalid transform', () => {
     const invalidTransforms = [
       '',
       'foo',
       'translateX(E2)',
+      'translateZ(10px)',
       'matrix(1, 2, foo, 4, 5, 6)',
       'matrix(1, 2)'
     ];
